# executor, planner: adjust the result of 'admin show ddl'