Improve performance to show onwer's address and the original query.

<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Adjust the result of 'admin show ddl'.

We'd better show owner address and the original query.

### What is changed and how it works?

Change the struct ShowDDLExec in executor/executor.go to show more information.


_before:_
```sql

mysql> admin show ddl;
+------------+--------------------------------------+--------------+--------------------------------------+
| SCHEMA_VER | OWNER                                | RUNNING_JOBS | SELF_ID                              |
+------------+--------------------------------------+--------------+--------------------------------------+
|       1011 | ee894c84-1309-4f98-8e59-0ca58709e426 |              | ee894c84-1309-4f98-8e59-0ca58709e426 |
+------------+--------------------------------------+--------------+--------------------------------------+
1 row in set (0.00 sec)

mysql> admin show ddl\G
*************************** 1. row ***************************
  SCHEMA_VER: 1057
       OWNER: ee894c84-1309-4f98-8e59-0ca58709e426
RUNNING_JOBS: 
     SELF_ID: ee894c84-1309-4f98-8e59-0ca58709e426
1 row in set (0.01 sec)


````
**after:**
```mysql
mysql> admin show ddl;
ERROR 2006 (HY000): MySQL server has gone away
No connection. Trying to reconnect...
Connection id:    2
Current database: *** NONE ***

+------------+--------------------------------------+---------------+--------------+--------------------------------------+-------+
| SCHEMA_VER | OWNER_ID                             | OWNER_ADDRESS | RUNNING_JOBS | SELF_ID                              | QUERY |
+------------+--------------------------------------+---------------+--------------+--------------------------------------+-------+
|        858 | 5449d580-074a-4efb-aeca-1c2e83f054ec | :4000         |              | 5449d580-074a-4efb-aeca-1c2e83f054ec |       |
+------------+--------------------------------------+---------------+--------------+--------------------------------------+-------+
1 row in set (0.00 sec)

mysql> admin show ddl\G
*************************** 1. row ***************************
   SCHEMA_VER: 858
     OWNER_ID: 5449d580-074a-4efb-aeca-1c2e83f054ec
OWNER_ADDRESS: :4000
 RUNNING_JOBS: 
      SELF_ID: 5449d580-074a-4efb-aeca-1c2e83f054ec
        QUERY: 
1 row in set (0.00 sec)

mysql> admin show ddl;
+------------+--------------------------------------+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------+-------------------------------------+
| SCHEMA_VER | OWNER_ID                             | OWNER_ADDRESS | RUNNING_JOBS                                                                                                                                                                                         | SELF_ID                              | QUERY                               |
+------------+--------------------------------------+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------+-------------------------------------+
|        947 | 5449d580-074a-4efb-aeca-1c2e83f054ec | :4000         | ID:1881, Type:create table, State:none, SchemaState:none, SchemaID:1700, TableID:1880, RowCount:0, ArgLen:0, start time: 2019-01-15 17:07:23.365 +0800 +08, Err:<nil>, ErrCount:0, SnapshotVersion:0 | 5449d580-074a-4efb-aeca-1c2e83f054ec | create table abcdhgife( ID int(4) ) |
+------------+--------------------------------------+---------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------+-------------------------------------+
1 row in set (0.01 sec)

mysql> admin show ddl\G
*************************** 1. row ***************************
   SCHEMA_VER: 984
     OWNER_ID: 5449d580-074a-4efb-aeca-1c2e83f054ec
OWNER_ADDRESS: :4000
 RUNNING_JOBS: ID:1955, Type:create table, State:none, SchemaState:none, SchemaID:1700, TableID:1954, RowCount:0, ArgLen:0, start time: 2019-01-15 17:07:25.059 +0800 +08, Err:<nil>, ErrCount:0, SnapshotVersion:0
      SELF_ID: 5449d580-074a-4efb-aeca-1c2e83f054ec
        QUERY: create table abcedgfhi( ID int(4) )
1 row in set (0.10 sec)

```