@zz-jason 2019-01-11
> > @xiekeyi98 Please add some proper labels for this PR, for example, component, type.
> 
> Sorry for my carelessness and newbie mistake.

It's OK, Don't worry about it. 😄