@xiekeyi98 2019-01-14
> Should we add an explanation in [this doc](https://github.com/pingcap/docs-cn/blob/master/sql/admin.md#admin-%E8%AF%AD%E5%8F%A5) for the result of admin show ddl?

In a way, I think the [doc](https://github.com/pingcap/docs-cn/blob/master/sql/admin.md#admin-%E8%AF%AD%E5%8F%A5) is enough now.