@zimulala 2019-01-16
/run-all-tests